requirements.txt
pandas
dataframe-image
matplotlib
flask
sqlalchemy
python-dotenv
psycopg2-binary
supabase
waitress
python-dotenv
fuzzywuzzy
python-Levenshtein
sentry-sdk
sentry-sdk[flask]
sentry-sdk[sqlalchemy]
apscheduler
requests
pytest
pytest-env